cc_fraud_detector_model/requirements.txt
pandas
matplotlib
numpy
seaborn
scipy
scikit-learn
pyyaml
python-box
ensure
xgboost